refactor: move `app` under `src`
