+++ src/app/main.py
@@ -2,8 +2,8 @@ import sys
 
 from PySide6 import QtGui, QtWidgets
 
-from app.ui.system_tray_icon import SystemTrayIcon
-from app.utils import get_asset
+from src.app.ui.system_tray_icon import SystemTrayIcon
+from src.app.utils import get_asset
 
 
 def main():
